+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- support Enum and Bitfield types with `ToJson` and `FromJson`.
+
 ### Changed
 - support for redscript v0.5.31.
+- bump `simdjson` to 4.6.3 (was 4.2.4).
 
 ------------------------